Move Gunicorn base applications to its own module
